Added Vector data structure.
The Vector data structure will be used to perform geometric
calculations that will lead us to our solution. The solution basically
being an addition of vectors from the starting point where the
vector to use in the next step of the algorithm is a function of
current facing direction (start North) and the Turn argument (from
input).