Added algorithm for determining first intersection.
- Modified Vector constructor to take in end point directly.
- Created points @property method on Vector which will generate the
points touched by the vector from the starting point to the end point.
- Added determineFirstIntersection algorithm to execute. The algorithm
adds the points touched by the vector created at each turn to a hash
map and ends on the first collision. We then take the vector created
by the point at that first collision and the start point to determine
the distance between the two.